fixed visual bug related to depth

+++ libs/sdw/CanvasPoint.cpp
@@ -3,21 +3,21 @@
 CanvasPoint::CanvasPoint() :
 		texturePoint(-1, -1) {}
 
-CanvasPoint::CanvasPoint(float xPos, float yPos) :
+CanvasPoint::CanvasPoint(int xPos, int yPos) :
 		x(xPos),
 		y(yPos),
 		depth(0.0),
 		brightness(1.0),
 		texturePoint(-1, -1) {}
 
-CanvasPoint::CanvasPoint(float xPos, float yPos, float pointDepth) :
+CanvasPoint::CanvasPoint(int xPos, int yPos, float pointDepth) :
 		x(xPos),
 		y(yPos),
 		depth(pointDepth),
 		brightness(1.0),
 		texturePoint(-1, -1) {}
 
-CanvasPoint::CanvasPoint(float xPos, float yPos, float pointDepth, float pointBrightness) :
+CanvasPoint::CanvasPoint(int xPos, int yPos, float pointDepth, float pointBrightness) :
 		x(xPos),
 		y(yPos),
 		depth(pointDepth),
